requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.7.2
attrs==25.4.0
beautifulsoup4==4.12.3
bitsandbytes==0.49.1
bitsandbytes-windows==0.37.5
blinker==1.9.0
blis==1.3.3
bs4==0.0.2
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
datasets==4.4.2
dill==0.4.0
diskcache==5.6.3
Django==4.2.6
duckdb==1.4.3
faiss-cpu==1.13.2
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
narwhals==2.15.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
plotly==6.5.1
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.2
pyarrow==22.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.3.1
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.2
setuptools==80.9.0
six==1.16.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.6
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.2
streamlit==1.53.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
Unidecode==1.4.0
urllib3==2.6.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
webencodings==0.5.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
